Some of this code is a bit hacky atm sorry :(.
Now there are two types of hazard:
- Tree in middle lane, must swerve around. Moving left or right logs time.
- Trees across all lanes, must brake. Braking and stopping logs time.

Collision vectors are now stored in an array of 100 (no way should we ever hit that).
The car won't accelerate while the brake is held down.

If Matt could maybe look at making it so that the car is stuck to the middle lane and "springs out" when the direction is held. Maybe only allow the car to change lane/swerve when a hazard is spawned in the middle lane.
Also I have only implemented keyboard controls, if someone could implement my changes for the gamepad that would be great.

Think that's it.